+++ src/orders/application/use-cases/deliver-order/deliver-order.command-handler.ts
@@ -23,7 +23,7 @@ export class DeliverOrderCommandHandler implements ICommandHandler<
 
     order.deliver();
 
-    if (!order.isValid()) {
+    if (!order.isValid) {
       const errors = order.brokenRules.getBrokenRules();
       throw new Error(errors.map((error) => error.message).join(', '));
     }