+++ src/orders/application/use-cases/remove-item-from-order/remove-item-from-order.command-handler.ts
@@ -25,7 +25,7 @@ export class RemoveItemFromOrderCommandHandler implements ICommandHandler<
     const productIdVO = IdValueObject.load(productId);
     order.removeItem(productIdVO);
 
-    if (!order.isValid()) {
+    if (!order.isValid) {
       const errors = order.brokenRules.getBrokenRules();
       throw new Error(errors.map((error) => error.message).join(', '));
     }