+++ src/orders/application/use-cases/ship-order/ship-order.command-handler.ts
@@ -23,7 +23,7 @@ export class ShipOrderCommandHandler implements ICommandHandler<
 
     order.ship(trackingNumber);
 
-    if (!order.isValid()) {
+    if (!order.isValid) {
       const errors = order.brokenRules.getBrokenRules();
       throw new Error(errors.map((error) => error.message).join(', '));
     }