+++ src/products/application/use-cases/create-product/create-product.command-handler.ts
@@ -23,7 +23,7 @@ export class CreateProductCommandHandler implements ICommandHandler<
       Price.create(price),
     );
 
-    if (!product.isValid()) {
+    if (!product.isValid) {
       const errors = product.brokenRules.getBrokenRules();
       throw new Error(errors.map((error) => error.message).join(', '));
     }